+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2023-04-17
+
+### Fixed
+
+#### Android
+
+- File upload on Android via webview doesn't work. Flutter team has to fix webview on their end.
+  Until then I used this small hack on Flutter side to able to upload photo from Android webview.
+    - Also I had to do the change on web side on 1.16.1 version
+    - Thanks to https://github.com/flutter/flutter/issues/118836
+
 ## [1.0.5] - 2023-04-13
 
 ### Changed